fix(mgcv-parity): PR #215 review round 1 -- correct provenance attribution, restore an independent test, validate inputs
Addresses the automated review's guardrail trip and two P1s (zero P0s,
zero test failures -- this was about the evidence record, not the maths).

[P1-1] The fixed-sp comparison IS independent (confirmed by the review's
own mechanical test on penalized_fit_and_score's signature), but nothing
declared a VerificationClaim for its actual producer -- the ledger and ADR
cited REML_SCORE_CLAIM, which covers a DIFFERENT fixture (2-block,
binomial-logit, paraPen-supplied via score_reml_point), not this session's
4-block binomial-cloglog formula-built one. Declares
gam_reml_optimize_conformance.FIXED_SP_MULTITERM_REML_CLAIM on the correct
producer, with a deviance companion mirroring REML_SCORE_CLAIM's own
scalePenalty-artifact rebuttal (added mgcv_deviance to the R probe's
payload). Corrects three now-contradictory "DIAGNOSTIC ONLY" comments
(the R probe header, two workflow step comments) written before ADR-210
fixed the criterion and never revisited.

[P1-2] gam_fixed_sp_score_compare.py now prints evidence_markdown() as its
headline instead of a hand-written table, and drops the stale "tighter
cut" correction column entirely (it was double-correcting an
already-fixed score) rather than patching it.

[P1-3], the guardrail trip: test_score_equals_the_explicit_dp_formula had
started calling observed_information_weight/logdet_s_plus directly instead
of independently re-deriving them. Split into a canonical-link case
(inlines the plain Fisher weight, valid only because alpha_i == 1 exactly
for binomial_logit -- a textbook identity) and a new non-canonical case
(binomial_cloglog, inlining Wood's own alpha_i formula from the paper,
with a sanity check that this fixture actually exercises alpha != 1).

P2s: reml_score_general now validates penalty_blocks/lambdas explicitly
(previously a bare IndexError/ValueError fired before the function's own
documented PolarisValidationError could); added missing dtype=np.float64
in three places; dropped a dead rng and parametrized a manual loop;
recorded the perf/history.jsonl creep verdict (no creep).

New test coverage: tests/test_analytics/test_gam_reml_optimize_conformance.py.

+++ src/polaris_re/analytics/gam_reml_appendix_b.py
@@ -290,8 +290,10 @@ def appendix_b_transform(
     # RESOLVED part only; pad with zero columns for any discarded true
     # null-space directions (q_accum's trailing q0-rank columns) so the
     # shapes line up with q_accum's full q0 columns.
-    e_square = scipy.linalg.block_diag(*e_blocks) if e_blocks else np.zeros((0, 0))
-    e_reduced = np.hstack([e_square, np.zeros((rank, q0 - rank))])
+    e_square = (
+        scipy.linalg.block_diag(*e_blocks) if e_blocks else np.zeros((0, 0), dtype=np.float64)
+    )
+    e_reduced = np.hstack([e_square, np.zeros((rank, q0 - rank), dtype=np.float64)])
     # Map back: q_accum @ (block-diagonal S in the resolved basis) @
     # q_accum^T equals the original (U_plus-projected) sum
     # `sum_i lambda_i * s_bar_i`, restricted to its positive eigenspace —
